+++ framework/views/slidingMenu.js
@@ -360,7 +360,7 @@ limitations under the License.
       },
 
       _appendMainPage: function(pageUrl, templateHTML) {
-        var pageScope = this._scope.$parent.$new();
+        var pageScope = this._scope.$new();
         var pageContent = angular.element(templateHTML);
         var link = $compile(pageContent);
 
@@ -382,7 +382,7 @@ limitations under the License.
        * @param {String}
        */
       _appendMenuPage: function(templateHTML) {
-        var pageScope = this._scope.$parent.$new();
+        var pageScope = this._scope.$new();
         var pageContent = angular.element(templateHTML);
         var link = $compile(pageContent);
 
@@ -594,8 +594,10 @@ limitations under the License.
       close: function(options) {
         options = options || {};
         options = typeof options == 'function' ? {callback: options} : options;
-       
-        this.emit('preclose');
+
+        this.emit('preclose', {
+          slidingMenu: this
+        });
 
         this._doorLock.waitUnlock(function() {
           this._logic.close(options);
@@ -613,7 +615,10 @@ limitations under the License.
           this._mainPage.children().css('pointer-events', '');
           this._mainPageHammer.off('tap', this._bindedOnTap);
 
-          this.emit('postclose');
+          this.emit('postclose', {
+            slidingMenu: this
+          });
+
           callback();
         }.bind(this), instant);
       },
@@ -638,7 +643,9 @@ limitations under the License.
         options = options || {};
         options = typeof options == 'function' ? {callback: options} : options;
 
-        this.emit('preopen');
+        this.emit('preopen', {
+          slidingMenu: this
+        });
 
         this._doorLock.waitUnlock(function() {
           this._logic.open(options);
@@ -656,7 +663,9 @@ limitations under the License.
           this._mainPage.children().css('pointer-events', 'none');
           this._mainPageHammer.on('tap', this._bindedOnTap);
 
-          this.emit('postopen');
+          this.emit('postopen', {
+            slidingMenu: this
+          });
 
           callback();
         }.bind(this), instant);